Initial commit
"""
This was very annoying and painful to make.
In the end, I went with a more 'hacky' route.
The code is a mess, hope you don't read it lol.

Anyways, the helmet item should work just fine.
Just note that it doesn't spawn yet, and I plan to add more when I have
more time.
-dastrukar
"""