Rework how armours are defined
No longer using hh_armourlist for defining armour, as that was way too
limited. (and i can't be arsed to make a parser and good syntax)
Now, armours are defined by creating a class that derives from
`HHArmourType` (Thinker)

HHArmourType has a function called DrawArmour, which is intended to
replace the default armour DrawHudStuff. (because i can't really
manipulate a function to not draw some numbers now can i)